poc/helper/app_transport_framework/message_to_send.py
from fhir.resources.R4B.bundle import Bundle


class MessageToSend:
    def __init__(self, atf_bundle: Bundle, receiver: str, message_type: str):
        self.atf_bundle = atf_bundle
        self.receiver = receiver
        self.message_type = message_type