poc/requirements.txt
fhir.resources
tzlocal
lxml
bs4
pdfkit